requirements.txt
docker
Django==4.1
crispy-bootstrap5
psycopg2
environs
django-q
djangorestframework